refactor: migrate types
