+++ packages/mask/src/plugins/RedPacket/SNSAdaptor/NftRedPacketHistoryItem.tsx
@@ -5,10 +5,9 @@ import { makeStyles, ActionButton } from '@masknet/theme'
 import { NetworkPluginID } from '@masknet/shared-base'
 import { isSameAddress, type NonFungibleCollection } from '@masknet/web3-shared-base'
 import { useChainContext, useNetworkDescriptor } from '@masknet/web3-hooks-base'
-import type { ChainId, SchemaType } from '@masknet/web3-shared-evm'
+import type { ChainId, SchemaType, NftRedPacketJSONPayload } from '@masknet/web3-shared-evm'
 import { Box, ListItem, Typography } from '@mui/material'
 import { dateTimeFormat } from '../../ITO/assets/formatDate.js'
-import type { NftRedPacketJSONPayload } from '../types.js'
 import { Translate, useI18N } from '../locales/index.js'
 import { useAvailabilityNftRedPacket } from './hooks/useAvailabilityNftRedPacket.js'
 import { useNftAvailabilityComputed } from './hooks/useNftAvailabilityComputed.js'
