+++ packages/mask/src/plugins/RedPacket/SNSAdaptor/RedPacket/index.tsx
@@ -1,7 +1,13 @@
 import { useCallback, useMemo, useState, useEffect } from 'react'
 import { Card, Typography, Box } from '@mui/material'
 import { Stack } from '@mui/system'
-import { ChainId, chainResolver, networkResolver } from '@masknet/web3-shared-evm'
+import {
+    ChainId,
+    chainResolver,
+    networkResolver,
+    RedPacketStatus,
+    type RedPacketJSONPayload,
+} from '@masknet/web3-shared-evm'
 import { formatBalance, isZero, TokenType } from '@masknet/web3-shared-base'
 import { NetworkPluginID } from '@masknet/shared-base'
 import { Web3 } from '@masknet/web3-providers'
@@ -14,8 +20,6 @@ import { isTwitter } from '../../../../social-network-adaptor/twitter.com/base.j
 import { useI18N as useBaseI18n } from '../../../../utils/index.js'
 import { useI18N } from '../../locales/index.js'
 import { LoadingBase, makeStyles, parseColor } from '@masknet/theme'
-import type { RedPacketJSONPayload } from '../../types.js'
-import { RedPacketStatus } from '../../types.js'
 import { useAvailabilityComputed } from '../hooks/useAvailabilityComputed.js'
 import { useClaimCallback } from '../hooks/useClaimCallback.js'
 import { useRefundCallback } from '../hooks/useRefundCallback.js'
