+++ packages/mask/src/plugins/RedPacket/SNSAdaptor/RedPacketDialog.tsx
@@ -10,7 +10,7 @@ import {
     useNetworkContext,
 } from '@masknet/web3-hooks-base'
 import { InjectedDialog, NetworkTab } from '@masknet/shared'
-import { ChainId, type GasConfig, GasEditor } from '@masknet/web3-shared-evm'
+import { ChainId, type GasConfig, GasEditor, type RedPacketJSONPayload } from '@masknet/web3-shared-evm'
 import { makeStyles, MaskTabList, useTabs } from '@masknet/theme'
 import { DialogContent, Tab } from '@mui/material'
 import { useActivatedPlugin } from '@masknet/plugin-infra/dom'
@@ -28,7 +28,7 @@ import { useI18N } from '../locales/index.js'
 import { useI18N as useBaseI18N } from '../../../utils/index.js'
 import { reduceUselessPayloadInfo } from './utils/reduceUselessPayloadInfo.js'
 import { RedPacketMetaKey } from '../constants.js'
-import { DialogTabs, type RedPacketJSONPayload } from '../types.js'
+import { DialogTabs } from '../types.js'
 import type { RedPacketSettings } from './hooks/useCreateCallback.js'
 import { RedPacketConfirmDialog } from './RedPacketConfirmDialog.js'
 import { RedPacketPast } from './RedPacketPast.js'
