+++ packages/mask/src/plugins/RedPacket/SNSAdaptor/RedPacketInHistoryList.tsx
@@ -2,12 +2,17 @@ import { type MouseEvent, useCallback, useState, useMemo, useRef, useEffect } fr
 import { useIntersectionObserver } from '@react-hookz/web'
 import { Box, Typography, Popper, useMediaQuery, type Theme, ListItem } from '@mui/material'
 import { makeStyles, ActionButton } from '@masknet/theme'
-import type { ChainId, SchemaType } from '@masknet/web3-shared-evm'
+import {
+    type ChainId,
+    type SchemaType,
+    type RedPacketJSONPayload,
+    type RedPacketJSONPayloadFromChain,
+    RedPacketStatus,
+} from '@masknet/web3-shared-evm'
 import intervalToDuration from 'date-fns/intervalToDuration'
 import nextDay from 'date-fns/nextDay'
 import { Translate, useI18N } from '../locales/index.js'
 import { dateTimeFormat } from '../../ITO/assets/formatDate.js'
-import { type RedPacketJSONPayload, type RedPacketJSONPayloadFromChain, RedPacketStatus } from '../types.js'
 import { useAvailabilityComputed } from './hooks/useAvailabilityComputed.js'
 import { useCreateRedPacketReceipt } from './hooks/useCreateRedPacketReceipt.js'
 import { useRefundCallback } from './hooks/useRefundCallback.js'
