+++ packages/mask/src/plugins/RedPacket/SNSAdaptor/RedPacketInPost.tsx
@@ -5,7 +5,7 @@ import { usePostInfoDetails } from '@masknet/plugin-infra/content-script'
 import { RedPacket } from './RedPacket/index.js'
 import { RedPacketRPC } from '../messages.js'
 import { activatedSocialNetworkUI } from '../../../social-network/index.js'
-import type { RedPacketJSONPayload, RedPacketRecord } from '../types.js'
+import type { RedPacketJSONPayload, RedPacketRecord } from '@masknet/web3-shared-evm'
 import { TransactionConfirmDialogProvider } from './context/TokenTransactionConfirmDialogContext.js'
 
 export interface RedPacketInPostProps {
