+++ packages/mask/src/plugins/RedPacket/SNSAdaptor/RedPacketNft.tsx
@@ -1,6 +1,6 @@
 import { useCallback, useEffect, useMemo, useState } from 'react'
 import { makeStyles, ActionButton, LoadingBase, parseColor, ShadowRootTooltip, useDetectOverflow } from '@masknet/theme'
-import { networkResolver, type ChainId } from '@masknet/web3-shared-evm'
+import { networkResolver, type ChainId, type RedPacketNftJSONPayload } from '@masknet/web3-shared-evm'
 import { Card, Typography, Button, Box } from '@mui/material'
 import { useTransactionConfirmDialog } from './context/TokenTransactionConfirmDialogContext.js'
 import { WalletConnectedBoundary, ChainBoundary, AssetPreviewer, NFTFallbackImage } from '@masknet/shared'
@@ -12,7 +12,6 @@ import { Icons } from '@masknet/icons'
 import { Stack } from '@mui/system'
 import { useI18N as useBaseI18N } from '../../../utils/index.js'
 import { useI18N } from '../locales/index.js'
-import type { RedPacketNftJSONPayload } from '../types.js'
 import { useClaimNftRedpacketCallback } from './hooks/useClaimNftRedpacketCallback.js'
 import { useAvailabilityNftRedPacket } from './hooks/useAvailabilityNftRedPacket.js'
 import { usePostLink } from '../../../components/DataSource/usePostInfo.js'
