+++ packages/mask/src/plugins/RedPacket/SNSAdaptor/RedPacketPast.tsx
@@ -3,12 +3,11 @@ import { TabPanel } from '@mui/lab'
 import { makeStyles } from '@masknet/theme'
 import { Box } from '@mui/material'
 import type { NonFungibleCollection } from '@masknet/web3-shared-base'
-import type { ChainId, SchemaType } from '@masknet/web3-shared-evm'
+import type { ChainId, SchemaType, RedPacketJSONPayload, NftRedPacketJSONPayload } from '@masknet/web3-shared-evm'
 import { useRemoteControlledDialog } from '@masknet/shared-base-ui'
 import { WalletMessages } from '@masknet/plugin-wallet'
 import { RedPacketHistoryList } from './RedPacketHistoryList.js'
 import { NftRedPacketHistoryList } from './NftRedPacketHistoryList.js'
-import type { RedPacketJSONPayload, NftRedPacketJSONPayload } from '../types.js'
 import { RedPacketNftMetaKey } from '../constants.js'
 import { useCurrentIdentity, useCurrentLinkedPersona } from '../../../components/DataSource/useActivatedUI.js'
 import { openComposition } from './openComposition.js'
