+++ packages/mask/src/plugins/RedPacket/SNSAdaptor/helpers.ts
@@ -1,6 +1,6 @@
 import { createRenderWithMetadata, createTypedMessageMetadataReader } from '@masknet/typed-message-react'
 import { RedPacketMetaKey, RedPacketNftMetaKey } from '../constants.js'
-import type { RedPacketJSONPayload, RedPacketNftJSONPayload } from '../types.js'
+import type { RedPacketJSONPayload, RedPacketNftJSONPayload } from '@masknet/web3-shared-evm'
 import schema from '../schema.json'
 import schemaNtf from '../schema-nft.json'
