+++ packages/mask/src/plugins/RedPacket/SNSAdaptor/index.tsx
@@ -11,7 +11,7 @@ import {
     renderWithRedPacketMetadata,
     renderWithRedPacketNftMetadata,
 } from './helpers.js'
-import type { RedPacketJSONPayload, RedPacketNftJSONPayload } from '../types.js'
+import type { RedPacketJSONPayload, RedPacketNftJSONPayload } from '@masknet/web3-shared-evm'
 import { RedPacketInjection } from './RedPacketInjection.js'
 import RedPacketDialog from './RedPacketDialog.js'
 import { RedPacketInPost } from './RedPacketInPost.js'
