+++ packages/mask/src/plugins/RedPacket/Worker/database.ts
@@ -1,5 +1,5 @@
 import { omit } from 'lodash-es'
-import type { RedPacketRecord, RedPacketRecordInDatabase, RedPacketNftRecordInDatabase } from '../types.js'
+import type { RedPacketRecord, RedPacketRecordInDatabase, RedPacketNftRecordInDatabase } from '@masknet/web3-shared-evm'
 import type { Plugin } from '@masknet/plugin-infra'
 
 export let RedPacketDatabase: Plugin.Worker.DatabaseStorage<RedPacketRecordInDatabase | RedPacketNftRecordInDatabase>
