+++ packages/web3-providers/src/RedPacket/constants.ts
@@ -2,3 +2,5 @@ export const TRANSACTIONS_BY_CONTRACT_METHOD_ENDPOINT =
     'https://arql3twjl4.execute-api.us-east-1.amazonaws.com/prod/records'
 
 export const MAX_SIZE_PER_PAGE = 1000
+
+export const CREATE_LUCKY_DROP_TOPIC = '0x86af556fd7cfab9462285ad44f2d5913527c539ff549f74731ca9997ca534018'
